agent/requirements.txt
web3==6.15.1
google-genai==1.0.0
python-dotenv==1.0.1